Added folder for unit tests and adjusted ci.yml
